#547 add missing file updates from #529
