+++ light-client/tests/supervisor.rs
@@ -121,8 +121,8 @@ fn run_multipeer_test(tc: TestBisection<AnonLightBlock>) {
 
 #[test]
 fn run_multipeer_tests() {
-    let mut tester = Tester::new(TEST_FILES_PATH);
+    let mut tester = Tester::new("bisection_multi_peer", TEST_FILES_PATH);
     tester.add_test("multipeer test", run_multipeer_test);
     tester.run_foreach_in_dir("bisection/multi_peer");
-    tester.print_results();
+    tester.finalize();
 }
